feat: product page add cart, folder organization
