+++ src/pages/categories/Index.js
@@ -32,7 +32,8 @@ import {
   IconAdjustmentsHorizontal,
   IconChevronDown,
 } from "@tabler/icons";
-import { FilterDrawer } from "../../components/filterDrawer/Index";
+// import { FilterDrawer } from "../../components/filterDrawer/Index";
+import { FilterDrawer } from "../../components/drawers/filterDrawer/Index";
 import { CartReducer } from "../../utils/cartReducer";
 import { useNavigate } from "react-router-dom";
